partial orbit db refactor
moving out all orbit db files into main src directory
