+++ src/orbit-db-console/src/db-config/index.js
