+++ src/orbit-db-console/src/db-views/index.js
